+++ bin/omen.js
@@ -1,10 +1,10 @@
 /*jslint node: true */
 "use strict";
 
-require('./../engine/utils/string_utils');
+require('./../engine/utils/general');
 
 var cli = require('cli').enable('status'),
-    commandUtils = require('./../engine/utils/command_utils');
+    commandUtils = require('./../engine/utils/commandUtils');
 
 GLOBAL.OMEN_CONFIG = require('./../config/app.json');
 
@@ -19,11 +19,11 @@ cli.main(function (args, options) {
     try {
         var command = args[0],
             filename = "project.json";
-        GLOBAL.OMEN_CLI_ARGS = args;
+        global.OMEN_CLI_ARGS = args;
 
-        GLOBAL.OMEN_ENV = GLOBAL.OMEN_CONFIG[GLOBAL.OMEN_CONFIG.env];
+        global.OMEN_ENV = global.OMEN_CONFIG[global.OMEN_CONFIG.env];
         if (process.env.OMEN_ENV)
-            GLOBAL.OMEN_ENV = GLOBAL.OMEN_CONFIG[process.env.OMEN_ENV];
+            global.OMEN_ENV = global.OMEN_CONFIG[process.env.OMEN_ENV];
 
         // Uncomment the following 3 lines to enable the possibility to select a file
         // from which to install dependencies.
@@ -32,15 +32,15 @@ cli.main(function (args, options) {
         //    command = args[1];
         //}
 
-        GLOBAL.OMEN_SAVE = options.save ? true : false;
-        GLOBAL.OMEN_ECLIPSE = options.eclipse ? true : false;
+        global.OMEN_SAVE = options.save ? true : false;
+        global.OMEN_ECLIPSE = options.eclipse ? true : false;
         
         commandUtils.SetInit(this, filename);
 
         if (options.version)
             return commandUtils.CommandExecutor('version').run();
 
-        return commandUtils.CommandExecutor(command).run(filename);
+        return commandUtils.CommandExecutor(command).run(global.OMEN_CLI_ARGS);
     } catch (err) {
         this.error(err.message);
         //throw err;