+++ engine/utils/general.js
@@ -35,3 +35,28 @@ if (typeof String.prototype.endsWith != 'function') {
         return this.slice(-str.length) == str;
     };
 }
+
+/**
+ * Object to hold general methods/functions.
+ *
+ * @class
+ * @return GeneralOmen
+ */
+var GeneralOmen = function () {
+    var self = this;
+
+    /**
+     * Tests to see if a given object is valid or not.
+     *
+     * @param {Object} object The object to be checked.
+     *
+     * @return boolean
+     */
+    self.isValid = function (object) {
+        return object !== null && object !== undefined;
+    };
+
+    return this;
+};
+
+global.GeneralOmen = GeneralOmen;